# src/app/api/v1/openapi.yaml
openapi: 3.0.3
info:
    title: AcidBot API v1
    version: 1.0.0
    description: |
        OpenAPI specification for the routes implemented under `src/app/api/v1`.

        Notes:
        - Several endpoints depend on an authenticated session in the application layer.
        - Image proxy responses return the upstream image bytes with the upstream content type.
        - Date fields are serialized by Next.js/MongoDB to ISO strings in JSON responses.
servers:
    - url: acid.kvznmx.com
      description: Production server
    - url: localhost:3000
      description: Local development server
...
paths:
    /api/v1/profile:
        get:
            tags: [profile]
            summary: Fetch the current maimai profile block
            description: |
                Fetches the maimai mobile profile page using the provided `clal` token,
                then parses the profile block into a normalized JSON object.

                This endpoint requires an authenticated session.
            parameters:
                - in: query
                  name: clal
                  required: true
                  schema:
                      type: string
                  description: The CLAL token used to fetch the maimai profile page.
            responses:
                '200':
                    description: Parsed profile information
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ParsedProfile'
                '401':
                    description: Unauthenticated
                '500':
                    description: Failed to fetch or parse the profile block
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /api/v1/profile/clal:
        get:
            tags: [profile]
            summary: Save a CLAL token for a user
            description: |
                Updates the user record with the provided `clal` token and redirects to a
                success or failure page. This endpoint is only a GET because the bookmarklet
                used is using a <script> tag, which cannot handle PUT.
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                      type: string
                  description: MongoDB ObjectId string for the user.
                - in: query
                  name: clal
                  required: true
                  schema:
                      type: string
                  description: CLAL token to persist.
            responses:
                '302':
                    description: Redirect to either `/pages/clal-fetch-success` or `/pages/clal-fetch-failure`
                    headers:
                        Location:
                            schema:
                                type: string
                            description: Redirect destination.
...
    /api/v1/sheets/level:
        post:
            tags: [sheets]
            summary: Fetch and normalize scores for a specific level page
            description: |
                Uses the provided CLAL token and redirect URL to fetch the level page,
                extracts score rows, joins them with song metadata, and returns computed
                rating data.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SheetsLevelRequest'
            responses:
                '200':
                    description: Normalized score list for the requested level
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/MSSB50'
                '500':
                    description: The page could not be fetched, or no valid results were found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /api/v1/b50:
        get:
            tags: [b50]
            summary: Fetch and store the current Best 50
            description: |
                Fetches the current B35/B15 results, enriches them with song metadata,
                persists the snapshot, and returns the computed Best 50 payload.

                This endpoint requires an authenticated session.
            parameters:
                - in: query
                  name: clal
                  required: true
                  schema:
                      type: string
                  description: The CLAL token used to fetch the maimai record pages.
            responses:
                '200':
                    description: Current Best 50 split into B35 and B15
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Best50Songs'
                '401':
                    description: Unauthenticated
                '500':
                    description: Failed to fetch, resolve, or persist the Best 50 payload
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
        post:
            tags: [b50]
            summary: Store an existing Best 50 snapshot
            description: |
                Saves the provided Best 50 payload to the historical collection with a
                derived total rating.

                This endpoint requires an authenticated session.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Best50Songs'
            responses:
                '200':
                    description: Snapshot stored successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: false
                                properties: {}
                '401':
                    description: Unauthenticated
                '500':
                    description: Failed to store the historical snapshot
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
...
components:
    schemas:
        ErrorResponse:
            type: object
            additionalProperties: true
            required: [error]
            properties:
                error:
                    oneOf:
                        - type: string
                        - type: object
                        - type: array
                        - type: number
                        - type: boolean
        UserCollectionCount:
            type: object
            additionalProperties: false
            required: [img, text]
            properties:
                img:
                    type: string
                    nullable: true
                text:
                    type: string
                    nullable: true
        ParsedProfile:
            type: object
            additionalProperties: false
            required:
                - profilePicture
                - dan
                - rank
                - userName
                - userDetail
                - userCollectionCount
            properties:
                profilePicture:
                    type: string
                    nullable: true
                    description: Same-origin proxied profile image URL when present.
                dan:
                    type: string
                    nullable: true
                    description: Same-origin proxied dan badge image URL when present.
                rank:
                    type: string
                    nullable: true
                    description: Same-origin proxied rank image URL when present.
                userName:
                    type: string
                    nullable: true
                userDetail:
                    type: string
                    nullable: true
                userCollectionCount:
                    nullable: true
                    oneOf:
                        - $ref: '#/components/schemas/UserCollectionCount'
        MaimaiSongScore:
            type: object
            additionalProperties: false
            required:
                - name
                - score
                - diff
                - dx
                - isDx
                - sync
                - combo
                - rank
            properties:
                name:
                    type: string
                score:
                    type: string
                diff:
                    type: string
                    enum: [basic, advanced, expert, master, remaster]
                dx:
                    type: string
                isDx:
                    type: string
                    enum: [dx, std]
                sync:
                    type: string
                    nullable: true
                    enum: [SYNC, FS, FS+, FDX, FDX+, null]
                combo:
                    type: string
                    nullable: true
                    enum: [FC, FC+, AP, AP+, null]
                rank:
                    type: string
                    nullable: true
                    enum:
                        [
                            D,
                            C,
                            B,
                            BB,
                            BBB,
                            A,
                            AA,
                            AAA,
                            S,
                            S+,
                            SS,
                            SS+,
                            SSS,
                            SSS+,
                            null,
                        ]
        MSSB50:
            allOf:
                - $ref: '#/components/schemas/MaimaiSongScore'
                - type: object
                  additionalProperties: false
                  required:
                      - levelConst
                      - rating
                      - dateIntlAdded
                      - achievement
                      - jacketURL
                  properties:
                      levelConst:
                          type: number
                      rating:
                          type: integer
                      dateIntlAdded:
                          type: string
                      achievement:
                          type: number
                      jacketURL:
                          type: string
        Best50Songs:
            type: object
            additionalProperties: false
            required: [b35, b15]
            properties:
                b35:
                    type: array
                    items:
                        $ref: '#/components/schemas/MSSB50'
                b15:
                    type: array
                    items:
                        $ref: '#/components/schemas/MSSB50'
...
        SheetsLevelRequest:
            type: object
            additionalProperties: false
            required: [clal, redirect]
            properties:
                clal:
                    type: string
                redirect:
                    type: string